Move FileExplorer component into a sub directory; Other small improvements.
